refactor: move arquivos para deixar mais explicito quais são do backend e quais são do front
